fix(web): keep WhatsApp phone parsing out of Node fs
Allowed numbers imported whatsapp-config, which loads node:fs/promises and crashed the Integrations page in the browser.
